renamed subpackage and updated some comments
