+++ src/probnum/__init__.py
@@ -19,7 +19,7 @@ quad
     Bayesian quadrature / numerical integration.
 random_variables
     Random variables representing uncertain values.
-testproblems
+problem_zoo
     Standard test problems for probabilistic numerical methods.
 utils
     Utility functions.
@@ -32,9 +32,9 @@ from . import (
     filtsmooth,
     linalg,
     linops,
+    problem_zoo,
     quad,
     random_variables,
-    testproblems,
     utils,
 )
 from .random_variables import RandomVariable, asrandvar
