+++ src/probnum/problem_zoo/linalg/_random_spd_matrix.py
@@ -18,10 +18,10 @@ def random_spd_matrix(
 
     Constructs a random symmetric positive definite matrix from a given spectrum. An
     orthogonal matrix :math:`Q` with :math:`\\operatorname{det}(Q)` (a rotation) is
-    sampled with respect to the Haar measure and the spectrum is rotated accordingly
-    resulting in :math:`A=Q \\operatorname{diag}(\\lambda_1, \\dots, \\lambda_n)
-    Q^\\top`. If no spectrum is provided, one is randomly drawn from a Gamma
-    distribution.
+    sampled with respect to the Haar measure and the diagonal matrix
+    containing the eigenvalues is rotated accordingly resulting in :math:`A=Q
+    \\operatorname{diag}(\\lambda_1, \\dots, \\lambda_n)Q^\\top`. If no spectrum is
+    provided, one is randomly drawn from a Gamma distribution.
 
     Parameters
     ----------
@@ -40,7 +40,7 @@ def random_spd_matrix(
 
     Examples
     --------
-    >>> from probnum.testproblems.linalg import random_spd_matrix
+    >>> from probnum.problem_zoo.linalg import random_spd_matrix
     >>> mat = random_spd_matrix(dim=5, random_state=0)
     >>> mat
     array([[10.49868572, -0.80840778,  0.79781892,  1.9229059 ,  0.73413367],
@@ -124,7 +124,7 @@ def random_sparse_spd_matrix(
 
     Examples
     --------
-    >>> from probnum.testproblems.linalg import random_sparse_spd_matrix
+    >>> from probnum.problem_zoo.linalg import random_sparse_spd_matrix
     >>> sparsemat = random_sparse_spd_matrix(dim=5, density=0.1, random_state=42)
     >>> sparsemat
     array([[1.        , 0.        , 0.        , 0.        , 0.        ],
