+++ tests/test_problem_zoo/test_linalg/test_random_spd_matrix.py
@@ -4,7 +4,7 @@ import unittest
 
 import numpy as np
 
-from probnum.testproblems.linalg import random_sparse_spd_matrix, random_spd_matrix
+from probnum.problem_zoo.linalg import random_sparse_spd_matrix, random_spd_matrix
 from tests.testing import NumpyAssertions
